Learned context functions and prompts moved to own file and learned_context must now be passed to assistant explicitly
