+++ assistant/__init__.py
@@ -2,3 +2,4 @@ from .assistant import Assistant, AssistantResponse
 from .gpt_assistant import GPTAssistant
 from .claude_assistant import ClaudeAssistant
 from .perplexity_assistant import PerplexityAssistant
+from .context import extract_learned_context
